requirements.txt
youtube-transcript-api
feedparser
requests
patchright
python-dotenv
